Move the utils to the utils folder
